payment-service/requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
requests==2.31.0
python-decouple==3.8
celery
redis